выделение базового класса
