+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/tests.uml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Diagram>
+  <ID>JAVA</ID>
+  <OriginalElement />
+  <nodes>
+    <node x="0.0" y="356.0">ru.stqa.pft.addressbook.GroupCreationTests</node>
+    <node x="90.5" y="0.0">ru.stqa.pft.addressbook.TestBase</node>
+    <node x="245.0" y="345.0">ru.stqa.pft.addressbook.GroupDeletionTests</node>
+  </nodes>
+  <notes />
+  <edges>
+    <edge source="ru.stqa.pft.addressbook.GroupCreationTests" target="ru.stqa.pft.addressbook.TestBase">
+      <point x="56.25" y="-26.0" />
+      <point x="168.75" y="320.0" />
+      <point x="201.875" y="320.0" />
+      <point x="-37.125" y="147.5" />
+    </edge>
+    <edge source="ru.stqa.pft.addressbook.GroupDeletionTests" target="ru.stqa.pft.addressbook.TestBase">
+      <point x="-64.25" y="-37.0" />
+      <point x="309.25" y="320.0" />
+      <point x="276.125" y="320.0" />
+      <point x="37.125" y="147.5" />
+    </edge>
+  </edges>
+  <settings layout="Hierarchic Group" zoom="1.0" x="0.0" y="0.0" />
+  <SelectedNodes />
+  <Categories>
+    <Category>Fields</Category>
+    <Category>Constructors</Category>
+    <Category>Methods</Category>
+  </Categories>
+  <SCOPE>All</SCOPE>
+  <VISIBILITY>private</VISIBILITY>
+</Diagram>
+
